+++ branches/2.0/src/main/java/org/jumpmind/symmetric/service/impl/RouterService.java
@@ -217,7 +217,7 @@ public class RouterService extends AbstractService implements IRouterService {
         long ts = System.currentTimeMillis();
         final int dataIdIncrementBy = parameterService.getInt(ParameterConstants.DATA_ID_INCREMENT_BY);
         long lastDataId = (Long) jdbcTemplate.query(getSql("selectDistinctDataIdFromDataEventSql"),
-                new Object[] { ref.getRefDataId() }, new int[] { Types.INTEGER },
+                new Object[] { ref.getRefDataId() }, new int[] { Types.NUMERIC },
                 new ResultSetExtractor<Long>() {
                     public Long extractData(ResultSet rs) throws SQLException, DataAccessException {
                         long lastDataId = ref.getRefDataId();
